requirements.txt
# Core Web Framework
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0

# Machine Learning
joblib==1.4.2

# Network Packet Analysis
scapy==2.5.0

# Note: scikit-learn/pandas/numpy are intentionally omitted to keep the
# web app runnable on environments where those wheels aren't available yet.
